+++ cli/commands_group.go
@@ -17,7 +17,7 @@ var Create = &ishell.Cmd{
 		req.Users = fnGetInputUser(c)
 		reqBytes, _ := req.Marshal()
 		reqDelegate := new(RequestDelegate)
-		if reqID, err := _SDK.ExecuteCommand(int64(msg.C_MessagesCreateGroup), reqBytes, reqDelegate, false); err != nil {
+		if reqID, err := _SDK.ExecuteCommand(msg.C_MessagesCreateGroup, reqBytes, reqDelegate, false); err != nil {
 			_Log.Debug(err.Error())
 		} else {
 			reqDelegate.RequestID = reqID
